summary: use pydantic for fetching the request args
The main point here is the post-process in the executions summary.
This is because the storage-manager doesn't really work with
summarizing status_display, which is a hybrid_property here.

So, I just wanted to add a hack here, to just postprocess the
status_display, to use status instead, and then rewrite the values
as needed.

However, of course the functions used request.args directly, so those
need to be moved into a separate "parsing", in a decorator in this
case, and using a pydantic model.