RND-495 Make managers.public_ip not unique
It's common - and somewhat expected - that multiple manager
(restservice) instances will be sharing the public-ip, when they're
behind a single LB/ingress/proxy.

In the future, it'd probably be better if there was only a single
entry in the db, perhaps? But anyway, while we do have multiple
managers that do indeed all respond behind the same public-ip,
that needs to be representable in the db...