RD-7154 blueprint upload: handle visibility correctly
When doing .to_dict(flat=False), all querystring arguments are going
to be lists. That's not what we want for essentially all of the
parameters available, but most notably, it breaks with visibility
(e.g. in the getting-started flow).

Instead, don't use flat=False, but keep the default of flat=True,
and get the list in the one and only argument that cares about getting
a list: the labels.