Add a timeout for listening to auditlog stream
During snapshot creation we don't want an open HTTP connection to be
hanging indefinitely.  This is why an auditlog listener will keep the
connection opened only for specified period of time.  If its thread was
not requested to stop, the connection will be re-established (for the
next period defined by the timeout parameter).