RND-364 Mark related entites for append (#4207)
* Add a timeout for listening to auditlog stream

During snapshot creation we don't want an open HTTP connection to be
hanging indefinitely.  This is why an auditlog listener will keep the
connection opened only for specified period of time.  If its thread was
not requested to stop, the connection will be re-established (for the
next period defined by the timeout parameter).

* Add mocked calls to auditlog.stream in snspshot tests

This prevents exceptions like this "object Mock can't be used in
'await' expression" in tests.

* Clean-up directory creation for *_events snapshots

* Pass the identifiers of snapshotted entities to the listener

Listener waits for (some) entities before starting to listen to
audit_log.

* Update tests

* Fix test_dont_append_new_blueprints audit_listener test

* Mark entities _related_to_snapshot for inclusion into the snapshot

* Test if executions related to dumped entities are going to be appended

* Pin urllib3 in amqp-postgres tests

* Fix type annotation: `dict[str, Any]` not `dict[str: Any]`