Store `latest_timestamp` of snapshotted entities
The value of this metadata is the maximum value of `created_at` (most of
the time), `uploaded_at` (plugins) or `timestamp` (events/logs) field of
the snapshotted entities.  There are some tables which do not store any
information about creation time (e.g. nodes, tenants etc.), in which
case `latest_timestamp` will not be stored.