RND-620 Store `latest_timestamp` for each entity type (#4214)
* Create the directories only for non-empty entities

Do not create the entities' directories prematurely, as there could be
none entities of given type.

* Store `latest_timestamp` of snapshotted entities

The value of this metadata is the maximum value of `created_at` (most of
the time), `uploaded_at` (plugins) or `timestamp` (events/logs) field of
the snapshotted entities.  There are some tables which do not store any
information about creation time (e.g. nodes, tenants etc.), in which
case `latest_timestamp` will not be stored.

* Store dumps as deployment_1.json, not just deployment_1 files

(this is a bugfix)

* Tests