Overwrite blueprint/deployment/plugin archives
In case an entity is appended to the snapshot (added again because of
the changes to it, reported by auditlog listener), let's just overwrite
the blueprint/deployment/plugin's archive.  This is safe to do because
ultimately we want to restore the final version of the entity, not the
intermediate ones.