Merge all the snapshot's entities before recreating them
This is a way in which we restore just the final entity's version.
Because of the auditlog listener and a possibility for the dump of an
entity to exist in multiple versions, let's build this giant dictionary
representing tenant-based part of the snapshot, which will keep just the
last entity's dump.

For most of the entities it's possible to find a unique identifier
(`id` for most of time, sometimes `name` or `username`).  For other
(`permissions` and `events`) we will build an artificial one.