RND-715 Zero-downtime snapshots tests (#4230)
* Overwrite blueprint/deployment/plugin archives

In case an entity is appended to the snapshot (added again because of
the changes to it, reported by auditlog listener), let's just overwrite
the blueprint/deployment/plugin's archive.  This is safe to do because
ultimately we want to restore the final version of the entity, not the
intermediate ones.

* Silly renames

* Merge all the snapshot's entities before recreating them

This is a way in which we restore just the final entity's version.
Because of the auditlog listener and a possibility for the dump of an
entity to exist in multiple versions, let's build this giant dictionary
representing tenant-based part of the snapshot, which will keep just the
last entity's dump.

For most of the entities it's possible to find a unique identifier
(`id` for most of time, sometimes `name` or `username`).  For other
(`permissions` and `events`) we will build an artificial one.

* Integration test for appending blueprints/deployments to a snapshot