RND-1091 node-instance update: skip version check for no-ops (#4263)
* Node instance update: skip the version check for no-op updates

The point of this commit is, if the update is not actually changing
anything, i.e. if the old runtime props are the same as new, skip
the version check, so that if many operations at the same tame set the
same values, they aren't denied.

Or in cases where runtime-props are "changed" to the same value,
which can still happen, and often happens in the case of e.g. the
terraform plugin.

* add a test

changing runtime props to the same value, even with the version being
older, doesn't actually throw a 409

* add an integration-test

This is actually the scenario from RND-1091, and let's check that
no conflict error is thrown