RND-1091 node-instance update: keep the version check
This partially reverts #4263 - the version check is done always, the
refactor stays...

Turns out, skipping the check doesn't quite work, but it breaks
operations when there's many concurrent runs. It seems that it's
important to keep the node-instances up to date on the workers.

I don't know the exact reason, but skipping this change, does make
operations start to overwrite each others' changes, which makes
the scale inte-tests flaky.