use boto3 for fileserver access (s3 api)
Instead of the requests calls, use boto3. The get/list/delete calls
are trivial and self-explanatory, but the proxy call is interesting.

We'll use a `proxy_pass $header_value` on the nginx side, so that
nginx proxies the user to the remote resource (on aws, or local on-prem
compatible storage). In order to do that, generate a pre-signed url
and put that in the header.