RND-1188 Use boto3 for s3 compatibility (#4277)
* Install boto3

* use boto3 for fileserver access (s3 api)

Instead of the requests calls, use boto3. The get/list/delete calls
are trivial and self-explanatory, but the proxy call is interesting.

We'll use a `proxy_pass $header_value` on the nginx side, so that
nginx proxies the user to the remote resource (on aws, or local on-prem
compatible storage). In order to do that, generate a pre-signed url
and put that in the header.

* removing files: also remove from the fileserver

removing blueprints, plugins, will now also remove them from the
fileserver, not just from the local filesystem

* flake

* init storage handler in the tests

* try to make snyk happy

* don't attempt to delete items that are already gone

* fix import

* fix the "is already gone" check

Use the full path, not just the fragment