CYBL-2089 migrate_pickle_to_json: use sqlalchemy core, not ORM
Instead of re-setting the parameters using the ORM, emit direct
UPDATE queries (via sqlalchemy core).

This way it will be much faster, but the point is, it'll not fire
any setter functions.

In this case, the setter we DON'T want to fire, is execution's
parameter re-evaluation (`merge_workflow_parameters`), because
that'll run the whole intrinsic function unnecessarily, and that
can fail.

Instead, we want to migrate data directly, just set whatever was
set before.