CYBL-2097 Snapshot creation/restore fixup (#4294)
* Implement upload snapshot archive as "upsert"

In case a snapshot already exists in a database (e.g. it is being
created in create_snapshot workflow), only update its archive.

* Upload snapshot archive after creation

In a distributed environment files created by mgmtworker must be
uploaded to "fileserver", which, in case of `file_server_type="local"`,
might be just a manager.

* Download snapshot archive before restoring

In a distributed environment uploaded snapshots are stored by the
"fileserver", which, in case of `file_server_type="local"`, might be
just a manager.  In any case, mgmtworker must fetch the archive it is
going to restore.