NE-6713 Avoid reloading config unnecessarily
When loading config, default the last_updated (if not set in the db,
which it won't be after a clean installation) to the current time.

Otherwise, we'll always consider last_updated to be None, which means
we'll reload the config every time...

Note: the date needs to be a string for this comparison, hence the
isoformat.