NE-6889 set-visibility under a transaction (#4309)
* NE-6889 set-visibility under a transaction

sm.update will try to remove items (if they're not part of the current
session cache) if they're duplicating tenant resources, and that
includes resources that were just set to global, if their id is
duplicated in another tenant.

Long story short, setting visibility to global, not only threw a 4xx
in case of conflict, but also deleted your resource. Funny.

Do that under a transaction so that doesn't happen. The resource is
considered persisted, and nothing actually happens.

This is the same thing done for all 4 kinds of set-visibility (note,
deployments already do run under a transaction for that), and
a test for each.

* fix test