NE-7791 configure_manager: use hostname for the admin token
This is going to be more pleasant to work with. In the k8s deployment
we'll be able to use a statefulset, with repeatable hostnames,
which will make it so we're not leaking tokens.

On aio, the name is going to be always the same then, which is
also going to make it so we're not leaking random tokens