NE-7791 Container scaling support (#4344)
* NE-7791 dockerfile: parametrize postgres host in entrypoints (#4340)

In the entrypoint, stop hardcoding the postgresql host as
`postgresql`. Instead, use the envvar

* NE-7791 configure_manager: use hostname for the admin token (#4341)

This is going to be more pleasant to work with. In the k8s deployment
we'll be able to use a statefulset, with repeatable hostnames,
which will make it so we're not leaking tokens.

On aio, the name is going to be always the same then, which is
also going to make it so we're not leaking random tokens