NE-16615 mgmtworker: parametrize wf/op workers separately (#248)
In this, we keep the default max_workers at 10 for workflows, but
now we have a separate amount for operations.

Note that the operations count should be larger than the workflows
count in order to avoid deadlocks: notably, with Components,
it's common that an operation on component runs a workflow on
that component's deployment.

So for example, if we have 10 components, we have 10 operations,
each runs a workflow, so 10 workflows, and... none of those workflows
can run operations of their own, because there's no operation workers
left! So we're stuck waiting forever in a deadlock.

If there's more operation workers than workflow workers, then there's
going to be a free slot for those workflows' operations.

Additionally, with this change, all of the mgmtworker command line
parameters will be able to be passed via envvars as well, which is
making it easier to configure them from a k8s deployment.
(command line still takes precedence over env)