NE-32415 storage_manager: use joinedload for associationproxy
The main change here is that in _resolve_included_fields, we now
use db.jopinedload rather than outerjoin. This populates the
assoc-proxy correctly, so that there's no longer 1+N queries when
listing deployments including the blueprint_id.

And so, other places that do require the attribute to e.g. sort on it,
do need an outerjoin, because assocproxies cant be used for sorting.

Also, rewrite the summarize function. This is the only place
that does grouping or with_entities, so let's avoid it polluting
the _get_query func