deployment to_response: careful about field names
we don't want to name fields in response_fields like
latest_execution when we only really want the id,
because that will cause the whole execution object
to be loaded.

To keep back-compat (ie. still return only the id
in the response, in a field named "latest_execution")
only load the id, but just rename the field back
in to_response