deployment: make deployment_groups be an assocproxy
This means we have double the same assocproxy, but oh well,
let's keep compat.

Anyway, we have to declare the "deployment_groups" as an available
include field indeed, but we dont want to put the whole objects
there, just ids.

So, the relationship for whole objects needs to be... hidden
somewhat.