NE-31415 storage_manager: use joinedload for associationproxy (#4445)
* NE-32415 storage_manager: use joinedload for associationproxy

The main change here is that in _resolve_included_fields, we now
use db.jopinedload rather than outerjoin. This populates the
assoc-proxy correctly, so that there's no longer 1+N queries when
listing deployments including the blueprint_id.

And so, other places that do require the attribute to e.g. sort on it,
do need an outerjoin, because assocproxies cant be used for sorting.

Also, rewrite the summarize function. This is the only place
that does grouping or with_entities, so let's avoid it polluting
the _get_query func

* when joinedloading, load_only the assocproxy field

dont load all the other fields unnecessarily

* dont include workflows in deployments list

Include workflows in deployments single-get, but not in list.
They're unnecessary in list, and computing the availability
for a workflow can be very expensive (because it needs to
examine node instance state)

* deployment to_response: careful about field names

we don't want to name fields in response_fields like
latest_execution when we only really want the id,
because that will cause the whole execution object
to be loaded.

To keep back-compat (ie. still return only the id
in the response, in a field named "latest_execution")
only load the id, but just rename the field back
in to_response

* storage_manager: dont shadow the "field"

into resolved_fields, we have to put the original "field",
in order for filtering to actually work correctly.
Don't overwrite that name with the proxy subfield.

* deployment: make deployment_groups be an assocproxy

This means we have double the same assocproxy, but oh well,
let's keep compat.

Anyway, we have to declare the "deployment_groups" as an available
include field indeed, but we dont want to put the whole objects
there, just ids.

So, the relationship for whole objects needs to be... hidden
somewhat.

* flake

* revert the deployment_groups stuff

I can't make it work, and it's not THAT important.
Deployment group objects arent typically very big.